Location permission on android added reverse geo location added

+++ shared/src/commonMain/kotlin/com/codekan/weathers/data/api/model/WeatherResponse.kt
@@ -86,7 +86,7 @@ data class Sys (
     @SerialName("type"    ) var type    : Int?    = null,
     @SerialName("id"      ) var id      : Int?    = null,
     @SerialName("country" ) var country : String? = null,
-    @SerialName("sunrise" ) var sunrise : Int?    = null,
-    @SerialName("sunset"  ) var sunset  : Int?    = null
+    @SerialName("sunrise" ) var sunrise : Long?    = null,
+    @SerialName("sunset"  ) var sunset  : Long?    = null
 
 )
